feat(fsd): A2 E3 — config/audit/upstream/user を entities へ・types/index.ts 撤去 (#376)
型のみ移設（Strategy R・挙動保存）で残り全型を entities へ:
- entities/user           (User)
- entities/clear-settings (ClearSettings + BankAccount 同梱・sibling import 回避／hub Q2)
- entities/audit-event    (AuditEvent + AuditAction 表示 union)
- entities/upstream-invoice (UpstreamInvoice + UpstreamClient 同梱・下記)

consumer 再ポイント: endpoints.ts の型 block・users/dunning/audit/settings page。
これで @/types 参照は 0 になり、空化した src/types/index.ts を削除。

【設計からの逸脱・要 hub 確認】UpstreamClient は実測で consumer 皆無の dead
export（移設前も types/index.ts 内のみ）。独立 upstream-client slice にすると
knip が dead file 検出で赤になるため、原ファイルで同一コメントに束ねられていた
upstream read models として entities/upstream-invoice に同梱（削除せず温存・
可逆）。分離 slice 化 or 削除の希望あれば調整する。

logic-diff-0 証明: 移設 7 型の body は md5 一致・types/index.ts は全行削除。
新規行は header doc comment のみ。npm run check 緑（vitest 62 不変）。

Refs #376
