+++ frontend/src/entities/audit-event/index.ts
@@ -1,0 +1,1 @@
+export type { AuditEvent, AuditAction } from './model'
