+++ frontend/src/entities/clear-settings/index.ts
@@ -1,0 +1,1 @@
+export type { ClearSettings, BankAccount } from './model'
