+++ frontend/src/entities/upstream-invoice/index.ts
@@ -1,0 +1,1 @@
+export type { UpstreamInvoice, UpstreamClient } from './model'
