+++ frontend/src/pages/settings/SettingsPage.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { getClearSettings, updateClearSettings, testUpstreamConnection } from '@/shared/api/endpoints'
-import type { BankAccount, ClearSettings } from '@/types'
+import type { BankAccount, ClearSettings } from '@/entities/clear-settings'
 import { Button } from '@/shared/ui/button'
 import { Card, CardFoot } from '@/shared/ui/card'
 import { Icon } from '@/shared/ui/icon'
