+++ frontend/src/pages/users/UsersPage.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { listUsers, createUser, deleteUser } from '@/shared/api/endpoints'
-import type { User } from '@/types'
+import type { User } from '@/entities/user'
 import { Button } from '@/shared/ui/button'
 import { Card } from '@/shared/ui/card'
 import { DataTable, TableStateRow } from '@/shared/ui/data-table'
